backend/requirements.txt
fastapi
uvicorn
pydantic
gunicorn
docker
python-dotenv
pandas
scikit-learn
joblib
docker==7.1.0
psycopg2-binary==2.9.9
dnspython==2.6.1
python-multipart
sqlalchemy
pydantic-settings
alembic
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.7.1
email-validator==2.1.1
PyJWT
alembic==1.13.1
urllib3==1.26.18
